backend/requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-RESTful==0.3.10
frozendict==2.4.4
html5lib==1.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
multitasking==0.0.11
niftystocks==0.1.2
numpy==2.1.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.4
yfinance==0.2.43
